Add update expression in breaking branch in a loop.
When replacing a for loop with while loop, typically
we insert the update statement as the last statement
in the loop body. However, if there are control
breaking statement (i.e. continue, break, return),
those branches would miss this update statement.
To solve this, we find out the breaking statemnts in
a loop body. Then we insert the update statement
immediately before the breaking statement.